Azure Token Creation Credentials Added
In order to create high level features within Azure an "AccessToken" has to be created to allow it (e.g. Resource Groups, Storagte Accounts, etc'.)

An additional static class in the Helper > Services folder that has been created but needs initiating and verification that it is prducing a valid token.