Resource Groups & Storage Account Sync with Azure
When creating a new company a new 'Resource Group" is automatically created with a compliant name.

When creating a new project and a new 'Storage Account' is automatically created with a compliant name and within the correct 'Resource Group'.

This applies to updating company and project names too. Modifications to the database have been made to track this information.

Delete for both has also been included.